refactor(link): add src, types, and tests folders
