+++ packages/link/package.json
@@ -4,6 +4,7 @@
   "author": "Tyson Warner <tyson.warner@availity.com>",
   "description": "React component that renders an a tag with the `href` formatted to leverage loadApp",
   "main": "index.js",
+  "types": "index.d.ts",
   "keywords": [
     "react",
     "availity"
@@ -20,7 +21,6 @@
     "access": "public"
   },
   "devDependencies": {
-    "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2"
   },
@@ -28,6 +28,7 @@
     "react": ">=16.3.0"
   },
   "dependencies": {
-    "@availity/resolve-url": "^1.1.21"
+    "@availity/resolve-url": "^1.1.21",
+    "prop-types": "^15.7.2"
   }
 }
