+++ packages/link/types/Link.d.ts
@@ -1,10 +1,10 @@
 import * as React from 'react';
 
 export interface AvLinkProps extends React.AnchorHTMLAttributes<HTMLAnchorElement> {
-    tag?: React.ReactType | string;
-    loadApp?: boolean;
+  tag?: React.ReactType | string;
+  loadApp?: boolean;
 }
 
-declare const AvLink: React.FunctionComponent<AvLinkProps>;
+declare const AvLink: React.FC<AvLinkProps>;
 
 export default AvLink;
